implement guess game retrieval in execute function with logging
